clickify server code and rename types to type
